Fix crash when drawing Boss Remains in the Get-Item state (Link holding object above head)
Instead of simply NOP-ing out the instructions in the Get-Item function which writes the object segmented address to the DList, calls hook function to check if the object is loaded into the game's object list and only writes the segmented address to the DList if so. Otherwise, assumes the caller wrote the segmented address to the DList.