Fix crash when loading room with world models loaded
Fixes a crash which would sometimes occur when loading a new room (in the same scene) with world models present. This would commonly occur when attempting to open doors in spider houses, as well as the door to Woodfall temple pre-boss room (with 4 stray fairies).

This crash was due to the way our object data heap worked for world models. Before loading a new room, it would clear the linear heap to allow for new object data to be loaded in. However, actors for the previous room would still be drawing, as well as actors for the new room. This behavior could lead to object data being relocated in the buffer within 1 frame, in a number of different ways. Due to the game having two display lists which it alternates between per frame, the pending display list would contain pointers to object data which had been overwritten, leading to a potential crash when processing that display list.

In order to address this, the linear heap (linheap) previously used to store object data has been updated to a "revolving" linear heap. This allows the heap to safely allocate and free regions without fully resetting the write pointer. Two new operations are added: "prepare" for advance, and "finish" advance. When "preparing", the most space-efficient pointer is determined and stored to be later used as the origin pointer. When "finishing", the origin pointer is updated. This means that when "preparing" (loading a new room), any previous or new object data which is allocated will still be located at the same address when "finished" (previous room is unloaded), during which data allocated before preparing is freed.

objheap is an abstraction layer built on top of linheap which is aware of object Ids and room indexes. It manages a list of object heap items and ensures that object data is being allocated and cleared properly for objects in relevant rooms.