Revert objheap if re-entering previous room during room transition
Implements "revert" operation for both linheap and objheap, to handle situations where the game prepares for a room transition (loads two rooms) then unloads the new room, which usually occurs when re-entering the previous room. This allows linheap and objheap to remain linear and safely free data for non-relevant rooms.