models.c: Add safeguard for objheap if loading new room before flushing heap operation
Adds a safeguard which prevents flushing an "advance" or "revert" operation if a new room is being loaded before the operation has been applied. Realistically this should never happen unless the player is able to cause a room load one frame after unloading a previous room?

If this happened without a safeguard: linheap would "prepare" while already in advance state (safe but inefficient), but "advance" or "revert" operation would then be flushed when not expected, invalidating the objheap state. For example, if a revert happens in this way, it could free data being actively used in the "next" room, which the heap could then overwrite and cause a crash.