Add subprojects MMR.Yaz and MMR.Yaz.CLI with the aim to provide performant Yaz0 encoding/decoding
This is an implementation of Yaz0 decoding & encoding which attempts to be performant and low-allocation. To achieve this, it makes use of more recent .NET constructs such as Span<T> and ref structs.

This implementation can perform both encode & decode operations without allocating anything on the heap. During encoding, it allows the user to optionally pass in buffers which are used for state, specifically for the "deltas window" and "recents." However, it can also allocate these buffers for the user on either the stack or the heap. If allocating these on the stack, both the decode & encode operations should not cause any heap allocations, with the possible exception of allocating the destination buffer if specified.

The terms "deltas window" and "recents" are specifically used during encoding and are a bit vague, so here is a brief explanation of what they are:

The "deltas window" refers to a buffer which stores deltas to previous indexes, where the value at these indexes matches the value of the current byte. This is used to efficiently "walk backwards" through previous indexes to look for matching strings during a lookup, operating similarly to a linked list.

The "recents" buffer is a mapping of byte values to the index of the most recent byte with that same value. This is used to find where in the "deltas window" the algorithm should start walking.

The subproject MMR.Yaz.CLI is less intended for general use and more for manual testing of the algorithm when needed.